backend/requirements.txt
djangorestframework
django
pygments
httpie